requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
greenlet==3.1.0
idna==3.10
imouapi==1.0.15
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mkdocs-material-extensions==1.3.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pip==24.2
pydocstyle==6.3.0
pyparsing==3.2.0
pyproject-api==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.35
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
yarl==1.11.1z 